requirements.txt
torch
transformers
pandas
safetensors
python-dotenv
#playwright
#playwright-stealth
emoji
accelerate